2409 Fix accept handling
To remove the dependency on autoneg we're replacing the parsing of the
accept header with some code that achieves nearly the same goal. The
code now allows the user to provide an extension to obtain a content-
type or to provide an accept-header and receive a content-type, an
indication of which renderer to use (soon to be removed hopefully)
and the name of the extension for that type.